requirements.txt
# =============================================================================
# Copilot-IA - Copilotos de Desenvolvimento com Agno
# =============================================================================

# Framework de Agentes IA
agno>=1.0.0
agno[openai]
agno[anthropic]
agno[groq]

# LLMs e Providers
openai>=1.40.0
anthropic>=0.34.0
google-generativeai>=0.7.0
groq>=0.9.0

# Processamento e Embeddings
tiktoken>=0.7.0
sentence-transformers>=3.0.0
chromadb>=0.5.0

# Framework Web e API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0

# Utilitários
python-dotenv>=1.0.0
httpx>=0.27.0
aiohttp>=3.10.0
tenacity>=8.5.0
rich>=13.7.0
typer>=0.12.0

# Análise de Código
tree-sitter>=0.22.0
tree-sitter-python>=0.22.0
tree-sitter-javascript>=0.21.0
tree-sitter-typescript>=0.21.0
pygments>=2.18.0
radon>=6.0.0
vulture>=2.11

# Git e Versionamento
gitpython>=3.1.43
pygithub>=2.3.0

# Cache e Storage
redis>=5.0.0
diskcache>=5.6.0

# Testes
pytest>=8.3.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0

# Desenvolvimento
jupyter>=1.0.0
ipykernel>=6.29.0
black>=24.4.0
ruff>=0.5.0
mypy>=1.11.0

# Logging e Monitoramento
structlog>=24.4.0
loguru>=0.7.0

